backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=22.1
psycopg2-binary>=2.9.3
python-dotenv>=1.0.0
Pillow>=9.0.0
requests>=2.31.0
django-extensions>=3.2.0
